refactor(hugr-py): rename extension to ext
